remove tx-head from ledger and use tx-address instead of tx-id
Since the db-head now has the tx-address in it we no longer need a separate tx-head. As
a matter of fact that's been obsolete since the transactor started tracking its own
head.

Addresses are to be used instead of ids anytime a reference to another component is
needed, so that data can be located over a network if necessary. Plus, they are a
superset of an id.